requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
build==1.4.0
certifi==2026.2.25
click==8.3.1
distro==1.9.0
Django==6.0.3
djangorestframework==3.16.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
openai==2.26.0
packaging==26.0
pip-tools==7.5.3
pydantic==2.12.5
pydantic_core==2.41.5
pyproject_hooks==1.2.0
setuptools==82.0.1
sniffio==1.3.1
sqlparse==0.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
wheel==0.46.3
